enable 1/2, 3/4, and 3/2 clock tournaments! - closes #1210

+++ modules/tournament/src/main/TournamentApi.scala
@@ -42,7 +42,7 @@ private[tournament] final class TournamentApi(
     var variant = chess.variant.Variant orDefault setup.variant
     val tour = Tournament.make(
       createdBy = me,
-      clock = TournamentClock(setup.clockTime * 60, setup.clockIncrement),
+      clock = TournamentClock((setup.clockTime * 60).toInt, setup.clockIncrement),
       minutes = setup.minutes,
       waitMinutes = setup.waitMinutes,
       mode = setup.mode.fold(Mode.default)(Mode.orDefault),
